perf vendor events: fix error code in json_events()
When EXPECT macro fails an assertion, the error code is
not properly set after the first loop of tokens.

This is because err is set to the return value from func
function pointer call, which must be 0, and is never reset
for each loop. I assume that this was not the intention, so
now err is set appropriately in EXPECT macro itself.

In addition to this, the indention in EXPECT macro is
tidied. The current indention alludes that the 2 statements
following the if statement are in the body, which is not
true.

Signed-off-by: John Garry <john.garry@huawei.com>