perf vendor events: drop support for unused topic directories
Currently a topic subdirectory is supported in the
pmu-events dir, in the following sample structure:
/arch/platform/subtopic/mysubtopic.json

Upto 256 levels of topic subdirectories are supported. So this
means that JSONs may be located in a topic dir as well as the
platform dir.

This topic subdirectory causes problems if we want to add support
for a vendor dir in the pmu-events structure (in the form
arch/platform/vendor), in that we cannot differentiate between
a vendor dir and a topic dir.

Since the topic dir feature is not used, drop it so it
does not block adding vendor subdirectory support.

NOTE: There is a bug in the current code with how add_topic()
sets topic_level. Also a delimiter could be added in printing
the "topic" in get_topic().

Signed-off-by: John Garry <john.garry@huawei.com>